Update cross section for all 3 year MC
Can't believe I have to clean up this after we go into 5 months production!
MC cross section could be off by a factor of 3 for background and a factor of
100 for the signal!

Current the same cross section is used for 3 year of MC. It is not most
accurate, but we will start for this.

The missing dataset are listed as well, which we need to submit!

The 2016 and 2017 preprocess list are checked. 2018 is quite messy due to
Scarlet's error production! Need to understand where we are to catch up this.
What a fucking mess